+++ examples/pingpong/src/ping.rs
@@ -3,21 +3,10 @@
 
 // region:		--- modules
 use chrono::Local;
-use clap::Parser;
 use dimas::prelude::*;
 use std::time::Duration;
 use tracing::info;
 // endregion:	--- modules
-
-// region:		--- Clap
-#[derive(Parser, Debug)]
-#[command(author, version, about, long_about = None)]
-struct Args {
-	/// prefix
-	#[arg(short, long, value_parser, default_value_t = String::from("examples"))]
-	prefix: String,
-}
-// endregion:	--- Clap
 
 #[derive(Debug)]
 struct AgentProps {
@@ -58,14 +47,11 @@ async fn main() -> Result<(), DimasError> {
 	// a tracing subscriber writing logs
 	tracing_subscriber::fmt::init();
 
-	// parse arguments
-	let args = Args::parse();
-
 	// create & initialize agents properties
 	let properties = AgentProps { counter: 0 };
 
-	// create an agent with the properties and the prefix given by `args`
-	let mut agent = Agent::new_with_prefix(Config::default(), properties, &args.prefix);
+	// create an agent with the properties and the prefix 'examples'
+	let mut agent = Agent::new_with_prefix(Config::default(), properties, "examples");
 
 	// create publisher for topic "ping"
 	agent.publisher().msg_type("ping").add()?;